Adding IMEI check, session in assessment, and fix minor bugs

+++ app/src/main/java/com/dlhk/smartpresence/api/response/data/DataPresence.kt
@@ -7,7 +7,7 @@ data class DataPresence(
     val employeeName: String,
     val employeeNumber: String,
     val livePhoto: Any,
-    val presenceId: Int,
+    val presenceId: Long,
     val presenceStatus: String,
     val regionName: String,
     val roleName: String,
